requirements.txt
requests==2.31.0
python-dotenv==1.0.0
openai==1.12.0
anthropic==0.18.1
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.6.0
aiohttp==3.9.3
